refactor: separate auth from user in the interface layer
