+++ src/main/kotlin/com/sexysisters/tojserverv2/interfaces/auth/AuthApiController.kt
@@ -1,10 +1,10 @@
-package com.sexysisters.tojserverv2.interfaces.user
+package com.sexysisters.tojserverv2.interfaces.auth
 
 import com.sexysisters.tojserverv2.common.response.CommonResponse
 import com.sexysisters.tojserverv2.domain.user.service.AuthService
-import com.sexysisters.tojserverv2.interfaces.user.dto.UserDtoMapper
-import com.sexysisters.tojserverv2.interfaces.user.dto.UserRequest
-import com.sexysisters.tojserverv2.interfaces.user.dto.UserResponse
+import com.sexysisters.tojserverv2.interfaces.auth.dto.AuthDtoMapper
+import com.sexysisters.tojserverv2.interfaces.auth.dto.AuthRequest
+import com.sexysisters.tojserverv2.interfaces.auth.dto.AuthResponse
 import io.swagger.annotations.Api
 import io.swagger.annotations.ApiOperation
 import org.springframework.web.bind.annotation.DeleteMapping
@@ -21,15 +21,15 @@ import javax.validation.Valid
 @RequestMapping("api/v2/auth")
 class AuthApiController(
     private val authService: AuthService,
-    private val userDtoMapper: UserDtoMapper,
+    private val authDtoMapper: AuthDtoMapper,
 ) {
 
     @ApiOperation(value = "기본 로그인")
     @PostMapping
-    fun login(@RequestBody @Valid request: UserRequest.Login): CommonResponse<UserResponse.Token> {
-        val userCommand = userDtoMapper.of(request)
+    fun login(@RequestBody @Valid request: AuthRequest.Login): CommonResponse<AuthResponse.Token> {
+        val userCommand = authDtoMapper.of(request)
         val userInfo = authService.login(userCommand)
-        val response = userDtoMapper.of(userInfo)
+        val response = authDtoMapper.of(userInfo)
         return CommonResponse.success(response)
     }
 
@@ -41,23 +41,23 @@ class AuthApiController(
 
     @ApiOperation(value = "액세스 토큰 재발급")
     @PutMapping
-    fun getNewAccessToken(@RequestHeader(value = "Refresh-Token") refreshToken: String): CommonResponse<UserResponse.Token> {
+    fun getNewAccessToken(@RequestHeader(value = "Refresh-Token") refreshToken: String): CommonResponse<AuthResponse.Token> {
         val accessToken = authService.getNewAccessToken(refreshToken)
-        val response = UserResponse.Token(accessToken)
+        val response = AuthResponse.Token(accessToken)
         return CommonResponse.success(response)
     }
 
     @ApiOperation(value = "인증 코드 이메일 발송")
     @PostMapping("/code")
-    fun sendCode(@RequestBody @Valid request: UserRequest.SendCode) {
-        val userCommand = userDtoMapper.of(request)
+    fun sendCode(@RequestBody @Valid request: AuthRequest.SendCode) {
+        val userCommand = authDtoMapper.of(request)
         authService.sendCode(userCommand)
     }
 
     @ApiOperation(value = "발송된 인증 코드 검증")
     @DeleteMapping("/code")
-    fun authenticateCode(@RequestBody @Valid request: UserRequest.AuthenticateCode): CommonResponse<Boolean> {
-        val userCommand = userDtoMapper.of(request)
+    fun authenticateCode(@RequestBody @Valid request: AuthRequest.AuthenticateCode): CommonResponse<Boolean> {
+        val userCommand = authDtoMapper.of(request)
         val response = authService.authenticateCode(userCommand)
         return CommonResponse.success(response)
     }
