+++ src/main/kotlin/com/sexysisters/tojserverv2/interfaces/auth/dto/AuthResponse.kt
@@ -1,0 +1,9 @@
+package com.sexysisters.tojserverv2.interfaces.auth.dto
+
+class AuthResponse {
+
+    class Token(
+        val accessToken: String,
+        val refreshToken: String? = null,
+    )
+}
