+++ src/main/kotlin/com/sexysisters/tojserverv2/interfaces/user/dto/UserRequest.kt
@@ -15,14 +15,6 @@ class UserRequest {
         val nickname: String,
     )
 
-    data class Login(
-        @field:NotBlank(message = "email is empty")
-        val email: String,
-
-        @field:NotBlank(message = "password is empty")
-        val password: String,
-    )
-
     data class GoogleAuth(
         @field:NotBlank(message = "auth code is empty")
         val code: String,
@@ -39,18 +31,5 @@ class UserRequest {
 
         @field:NotBlank(message = "name is empty")
         val name: String,
-    )
-
-    data class SendCode(
-        @field:NotBlank(message = "email is empty")
-        val email: String,
-    )
-
-    data class AuthenticateCode(
-        @field:NotBlank(message = "email is empty")
-        val email: String,
-
-        @field:NotBlank(message = "code is empty")
-        val code: String,
     )
 }
